+++ packages/time/spec.js
@@ -12,6 +12,7 @@ describe('@routes/time', () => {
 		const app = express();
 		app.use(measure);
 		app.get('/user/:user_id', async(req, res, next) => {
+			res.set('Service-Name', 'my-service');
 			await wait(delay);
 			res.sendStatus(206);
 			next();
@@ -26,12 +27,14 @@ describe('@routes/time', () => {
 	it('measures the route', async() => {
 		await fetch(`http://127.0.0.1:${port}/user/234`);
 		const [ {
-			duration, method, route, status, url,
+			duration, method, route, status, request, response
 		} ] = mesaurements;
 		expect(duration).to.be.at.least(100);
 		expect(method).to.equal('GET');
 		expect(route).to.equal('/user/:user_id');
-		expect(url).to.equal('/user/234');
 		expect(status).to.equal(206);
+		expect(request.url).to.equal('/user/234');
+		console.log('ENDED', response.finished);
+		expect(response.get('service-name')).to.equal('my-service');
 	});
 });