+++ packages/time/CHANGELOG.md
@@ -2,7 +2,7 @@
 ## 2022-06-06
 - Replace "request.end" monkey-patch with on "finish" listener
 - Fix double slashes on router routes
-- Add URL argument to measurement
+- Add "request" and "response" to measurement arguments
 
 # 2.0.3
 ## 2021-10-13